requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Flask-Limiter==3.5.0